feat:新增地址

+++ src/main/java/com/darkcom/decoration/model/Address.java
@@ -3,7 +3,7 @@ package com.darkcom.decoration.model;
 import java.util.Date;
 
 public class Address {
-    private Long id;
+    private Integer id;
 
     private String phone;
 
@@ -25,13 +25,15 @@ public class Address {
 
     private String account;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
+
+
 
     public String getPhone() {
         return phone;
